refactor: remove redundant cast in parseHexOrDec32 range check
Agent-Logs-Url: https://github.com/haoyi-jason/ModbusScope/sessions/24eeb893-acba-4b1d-9efb-72aba09476c6

+++ src/dialogs/batchparamdialog.cpp
@@ -437,7 +437,7 @@ quint32 BatchParamDialog::parseHexOrDec32(const QString& text, bool* ok)
         const qint64 signed64 = trimmed.toLongLong(&localOk, 10);
         if (localOk)
         {
-            if (signed64 < static_cast<qint64>(INT32_MIN) || signed64 > static_cast<qint64>(UINT32_MAX))
+            if (signed64 < INT32_MIN || signed64 > static_cast<qint64>(UINT32_MAX))
             {
                 localOk = false;
             }
